いろいろ変更

+++ UsbHid/UsbHid.cpp
@@ -56,7 +56,7 @@ namespace
   }
 }
 
-void tud_hid_report_complete_cb(uint8_t instance, uint8_t const *report, uint8_t len)
+void tud_hid_report_complete_cb(uint8_t instance, uint8_t const *report, uint16_t len)
 {
   notify_report_complete();
 }
